Refactorización: se organizó el AR: Consultorio dentro de su propio paquete
