+++ domain/src/main/java/co/com/sofkau/clinica/administracion/consultorio/ConsultorioEventChange.java
@@ -1,9 +1,9 @@
-package co.com.sofkau.clinica.administracion.consultorio.commands;
+package co.com.sofkau.clinica.administracion.consultorio;
 
 import co.com.sofka.domain.generic.EventChange;
-import co.com.sofkau.clinica.administracion.consultorio.commands.events.AuxiliarAsignado;
-import co.com.sofkau.clinica.administracion.consultorio.commands.events.ConsultorioCreado;
-import co.com.sofkau.clinica.administracion.consultorio.commands.events.MedicoAsignado;
+import co.com.sofkau.clinica.administracion.consultorio.events.AuxiliarAsignado;
+import co.com.sofkau.clinica.administracion.consultorio.events.ConsultorioCreado;
+import co.com.sofkau.clinica.administracion.consultorio.events.MedicoAsignado;
 
 public class ConsultorioEventChange extends EventChange {
     public ConsultorioEventChange(Consultorio consultorio) {
