+++ domain/src/main/java/co/com/sofkau/clinica/administracion/consultorio/commands/AsignarAuxiliar.java
@@ -1,8 +1,8 @@
 package co.com.sofkau.clinica.administracion.consultorio.commands;
 
 import co.com.sofka.domain.generic.Command;
-import co.com.sofkau.clinica.administracion.consultorio.commands.values.AuxiliarId;
-import co.com.sofkau.clinica.administracion.consultorio.commands.values.Nombre;
+import co.com.sofkau.clinica.administracion.consultorio.values.AuxiliarId;
+import co.com.sofkau.clinica.administracion.consultorio.values.Nombre;
 
 public class AsignarAuxiliar extends Command {
