+++ domain/src/main/java/co/com/sofkau/clinica/administracion/consultorio/events/AuxiliarAsignado.java
@@ -1,8 +1,8 @@
-package co.com.sofkau.clinica.administracion.consultorio.commands.events;
+package co.com.sofkau.clinica.administracion.consultorio.events;
 
 import co.com.sofka.domain.generic.DomainEvent;
-import co.com.sofkau.clinica.administracion.consultorio.commands.values.AuxiliarId;
-import co.com.sofkau.clinica.administracion.consultorio.commands.values.Nombre;
+import co.com.sofkau.clinica.administracion.consultorio.values.AuxiliarId;
+import co.com.sofkau.clinica.administracion.consultorio.values.Nombre;
 
 public class AuxiliarAsignado extends DomainEvent {
